feat(Greeting): Composable Text function
