+++ app/src/main/java/com/example/learningtest/compose/basic/codelab/Greeting.kt
@@ -1,0 +1,26 @@
+package com.example.learningtest.compose.basic.codelab
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.example.learningtest.ui.theme.LearningTestTheme
+
+@Composable
+fun Greeting(
+    name: String,
+    modifier: Modifier = Modifier,
+) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier,
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun GreetingPreview() {
+    LearningTestTheme {
+        Greeting("Android")
+    }
+}
